chore: rename airdrop to faucet
