+++ faucet/.gitignore
