+++ faucet/api/pangolin/config/chain.json
