+++ faucet/api/utils.ts
